requirements.txt
flask
flask-cors
pytesseract
pillow
llama-cpp-python
pdf2image